IFS-6696 - Update to formValidation number to exclude step of decimal to stop the whole number validation appearing and step added to templates.

+++ ifs-web-service/ifs-project-setup-mgt-service/src/test/java/org/innovateuk/ifs/project/funding/level/controller/ProjectFinanceFundingLevelControllerTest.java
@@ -129,7 +129,7 @@ public class ProjectFinanceFundingLevelControllerTest extends BaseControllerMock
         verify(financeRowRestService).update(academicFinances.getGrantClaim());
         verify(financeRowRestService).update(industrialFinances.getGrantClaim());
 
-        assertEquals(60, academicFinances.getGrantClaimPercentage());
+        assertEquals(BigDecimal.valueOf(60), academicFinances.getGrantClaimPercentage());
         assertEquals(60, industrialFinances.getGrantClaimPercentage());
     }
